Update spritz.js
Fixed the split that was done in the pivot function for longer words
There seems to be a problem with longer words here. especially if special characters are hexed
This assumes that the max word size is 22-7
let us rather split the difference. 
If the word is longer than this 22 then just show it all